chore(replay): delete the TS scrubbers from the ml-mirror pipeline
The native Rust addon (@posthog/replay-anonymizer) has been the default
anonymization path since SESSION_RECORDING_ML_RUST_ANONYMIZER defaulted to
true, leaving the TypeScript scrubbers as dead fallback code. Remove them
and the flag; the ml-mirror pipeline now always uses the fused
parse+anonymize step, and the server unconditionally initializes and
health-checks the addon at startup.

firstPartyHostPatterns moves from anonymize/url.ts to
shared/teams/first-party-hosts.ts since the Rust path still consumes it via
the team service. The allow-list loading (allow-lists.ts,
allow-list-loader.ts, default-dict.ts) stays: the server feeds the loaded
lists into the addon.

Fixture coverage is unchanged: the TS-parity half of shared-fixtures.test.ts
goes away, while the addon-through-FFI half remains, and the pure-Rust side
keeps running via rust/replay-anonymizer-node/tests/parity.rs.

Generated-By: PostHog Code
Task-Id: 273554de-31f0-49a3-b3d6-f94d9c8fa669
