--- nodejs/src/ingestion/pipelines/sessionreplay/anonymize/index.ts
@@ -1,5 +1,0 @@
-export { AllowLists } from './allow-lists'
-export { anonymizeEvent, anonymizeParsedMessage } from './anonymize-event'
-export { loadAllowLists } from './allow-list-loader'
-export { type ScrubContext } from './config'
-export { defaultAllowLists } from './default-dict'
