+++ nodejs/src/ingestion/pipelines/sessionreplay/ml-mirror/ml-mirror-pipeline.test.ts
@@ -4,6 +4,8 @@ import { Message } from 'node-rdkafka'
 import { DLQ_OUTPUT, INGESTION_WARNINGS_OUTPUT, OVERFLOW_OUTPUT } from '~/common/outputs'
 import { IngestionOutputs } from '~/common/outputs/ingestion-outputs'
 import { EventIngestionRestrictionManager } from '~/common/utils/event-ingestion-restrictions'
+import { parseJSON } from '~/common/utils/json-parse'
+import { logger } from '~/common/utils/logger'
 import { PromiseScheduler } from '~/common/utils/promise-scheduler'
 import { createApplyEventRestrictionsStep, createParseHeadersStep } from '~/ingestion/common/steps/event-preprocessing'
 import { TopHogRegistry } from '~/ingestion/framework/extensions/tophog'
@@ -34,6 +36,17 @@ jest.mock('~/ingestion/common/steps/event-preprocessing', () => ({
 const mockCreateParseHeadersStep = createParseHeadersStep as jest.Mock
 const mockCreateApplyEventRestrictionsStep = createApplyEventRestrictionsStep as jest.Mock
 
+// The pipeline's parse+anonymize step runs inside the native addon; scrub-dependent tests need it built.
+let rustAddon: typeof import('@posthog/replay-anonymizer') | null = null
+try {
+    rustAddon = require('@posthog/replay-anonymizer')
+} catch (e) {
+    if (process.env.CI) {
+        throw new Error(`replay-anonymizer addon failed to load; pipeline tests cannot run in CI: ${String(e)}`)
+    }
+    logger.warn('🙈', 'replay_anonymizer_addon_not_built_skipping_pipeline_scrub_tests')
+}
+
 function createMockTopHog(): TopHogRegistry {
     const recorder = { record: jest.fn() }
     return {
@@ -52,7 +65,6 @@ describe('ml-mirror-pipeline', () => {
     let outputs: jest.Mocked<
         IngestionOutputs<typeof DLQ_OUTPUT | typeof OVERFLOW_OUTPUT | typeof INGESTION_WARNINGS_OUTPUT>
     >
-    const scrubContext = { allow: defaultAllowLists() }
 
     // Resolves every session to 30d so messages flow through to recording.
     const retentionService = {
@@ -136,7 +148,6 @@ describe('ml-mirror-pipeline', () => {
             topHog,
             sessionBatchManager: mockSessionBatchManager,
             isDebugLoggingEnabled: () => false,
-            scrubContext,
         })
     }
 
@@ -221,20 +232,6 @@ describe('ml-mirror-pipeline', () => {
         } as unknown as Message
     }
 
-    it('anonymizes events before recording for an opted-in team', async () => {
-        mockTeamService = {
-            getTeamByToken: jest.fn().mockResolvedValue(team(true)),
-            getRetentionPeriodByTeamId: jest.fn().mockResolvedValue(30),
-        } as unknown as TeamService
-
-        await runSessionReplayPipeline(buildPipeline(), [message('sess-1')])
-
-        expect(recordMock).toHaveBeenCalledTimes(1)
-        const recorded = recordMock.mock.calls[0][0]
-        // The Input event's text was scrubbed before it reached the recorder.
-        expect(recorded.message.eventsByWindowId['window-1'][0].data.text).toBe('Hello **********')
-    })
-
     it('drops sessions for a team that did not opt into AI training', async () => {
         mockTeamService = {
             getTeamByToken: jest.fn().mockResolvedValue(team(false)),
@@ -246,19 +243,51 @@ describe('ml-mirror-pipeline', () => {
         expect(recordMock).not.toHaveBeenCalled()
     })
 
-    it('scrubs a FullSnapshot (text, url, free-text data-*) end-to-end before recording', async () => {
-        mockTeamService = {
-            getTeamByToken: jest.fn().mockResolvedValue(team(true)),
-            getRetentionPeriodByTeamId: jest.fn().mockResolvedValue(30),
-        } as unknown as TeamService
+    const describeAddon = rustAddon ? describe : describe.skip
+    describeAddon('with the native anonymizer', () => {
+        beforeAll(() => {
+            rustAddon!.initAnonymizer(defaultAllowLists().entries())
+        })
 
-        await runSessionReplayPipeline(buildPipeline(), [fullSnapshotMessage('sess-3')])
+        // The fused step emits pre-serialized JSONL lines of [windowId, event].
+        function recordedEvents(): [string, any][] {
+            const lines: Buffer = recordMock.mock.calls[0][0].message.preSerialized.lines
+            return lines
+                .toString()
+                .split('\n')
+                .filter((l) => l.length > 0)
+                .map((l) => parseJSON(l))
+        }
 
-        expect(recordMock).toHaveBeenCalledTimes(1)
-        const node = recordMock.mock.calls[0][0].message.eventsByWindowId['window-1'][0].data.node.childNodes[0]
-        expect(node.childNodes[0].textContent).toBe('Hello **********') // DOM text
-        expect(node.attributes.href).toContain('https://example.com/') // authority kept...
-        expect(node.attributes.href).not.toContain('abc') // ...path segments redacted
-        expect(node.attributes['data-note']).not.toContain('Smithson') // free-text data-* scrubbed
+        it('anonymizes events before recording for an opted-in team', async () => {
+            mockTeamService = {
+                getTeamByToken: jest.fn().mockResolvedValue(team(true)),
+                getRetentionPeriodByTeamId: jest.fn().mockResolvedValue(30),
+            } as unknown as TeamService
+
+            await runSessionReplayPipeline(buildPipeline(), [message('sess-1')])
+
+            expect(recordMock).toHaveBeenCalledTimes(1)
+            const [windowId, event] = recordedEvents()[0]
+            expect(windowId).toBe('window-1')
+            // The Input event's text was scrubbed before it reached the recorder.
+            expect(event.data.text).toBe('Hello **********')
+        })
+
+        it('scrubs a FullSnapshot (text, url, free-text data-*) end-to-end before recording', async () => {
+            mockTeamService = {
+                getTeamByToken: jest.fn().mockResolvedValue(team(true)),
+                getRetentionPeriodByTeamId: jest.fn().mockResolvedValue(30),
+            } as unknown as TeamService
+
+            await runSessionReplayPipeline(buildPipeline(), [fullSnapshotMessage('sess-3')])
+
+            expect(recordMock).toHaveBeenCalledTimes(1)
+            const node = recordedEvents()[0][1].data.node.childNodes[0]
+            expect(node.childNodes[0].textContent).toBe('Hello **********') // DOM text
+            expect(node.attributes.href).toContain('https://example.com/') // authority kept...
+            expect(node.attributes.href).not.toContain('abc') // ...path segments redacted
+            expect(node.attributes['data-note']).not.toContain('Smithson') // free-text data-* scrubbed
+        })
     })
 })
